improve device configuration page to now display per device category connected devices

+++ BlazorCamPortal/Services/Services/Devices/DeviceTypeService.cs
@@ -3,6 +3,9 @@ using CamPortal.Contracts.Abstractions.Repositories;
 using CamPortal.Contracts.Abstractions.Services;
 using CamPortal.Contracts.Dtos.DeviceTypeDtos;
 using CamPortal.Contracts.Models;
+using CamPortal.Core.Utilities;
+using Microsoft.Extensions.Configuration;
+using System.ComponentModel.DataAnnotations;
 
 namespace CamPortal.Core.Services.Devices
 {
@@ -12,14 +15,20 @@ namespace CamPortal.Core.Services.Devices
         private readonly IDeviceTypeIconStorageService _deviceTypeIconStorageService;
         private readonly IMapper _mapper;
 
+        private readonly string _defaultIconName;
+
         public DeviceTypeService(
             IDeviceTypeRepository deviceTypeRepository,
             IDeviceTypeIconStorageService deviceTypeIconStorageService,
-            IMapper mapper)
+            IMapper mapper,
+            IConfiguration configuration)
         {
             _deviceTypeRepository = deviceTypeRepository;
             _deviceTypeIconStorageService = deviceTypeIconStorageService;
             _mapper = mapper;
+
+            _defaultIconName = configuration.GetSection("DeviceTypeIconsConfig")["DefaultIconName"]
+                ?? throw new ArgumentNullException("DeviceTypeIconsConfig:DefaultIconName not configured");
         }
 
         public async Task<List<DeviceTypeDisplayModel>> GetAllDeviceTypesAsync()
@@ -30,7 +39,6 @@ namespace CamPortal.Core.Services.Devices
                 .Select(dto =>
                 {
                     var model = _mapper.Map<DeviceTypeDisplayModel>(dto);
-                    model.IconUrl = _deviceTypeIconStorageService.BuildPublicUrl(model.IconName, model.IconUpdatedAt);
                     return model;
                 })
                 .ToList();
@@ -38,9 +46,9 @@ namespace CamPortal.Core.Services.Devices
 
         public async Task<Guid> CreateDeviceTypeAsync(CreateDeviceTypeModel model, CancellationToken ct)
         {
-            if (model.IconFile is null)
+            if (!MiscUtilities.ValidateModel(model, out ICollection<ValidationResult> validationResults))
             {
-                throw new InvalidOperationException("Icon file is required.");
+                throw new ArgumentException(string.Join(", ", validationResults.Where(x => !string.IsNullOrEmpty(x.ErrorMessage)).Select(v => v.ErrorMessage)));
             }
 
             if (await _deviceTypeRepository.DoesExistByNameAsync(model.Name))
@@ -48,7 +56,7 @@ namespace CamPortal.Core.Services.Devices
                 throw new InvalidOperationException($"A device type named '{model.Name}' already exists.");
             }
 
-            var iconName = await _deviceTypeIconStorageService.SaveAsync(model.IconFile, ct);
+            var iconName = model.IconFile == null ? _defaultIconName : await _deviceTypeIconStorageService.SaveAsync(model.IconFile, ct);
 
             var dto = new CreateDeviceTypeDto
             {
@@ -56,6 +64,7 @@ namespace CamPortal.Core.Services.Devices
                 IconName = iconName,
                 IconUpdatedAt = DateTime.UtcNow,
                 DeviceCategory = model.DeviceCategory,
+                Description = model.Description,
             };
 
             try
@@ -78,7 +87,7 @@ namespace CamPortal.Core.Services.Devices
             }
 
             var deleted = await _deviceTypeRepository.DeleteTypeAsync(deviceTypeId);
-            if (deleted)
+            if (deleted && string.Compare(dto.IconName, _defaultIconName, true) != 0)
             {
                 await _deviceTypeIconStorageService.DeleteAsync(dto.IconName);
             }
@@ -99,7 +108,6 @@ namespace CamPortal.Core.Services.Devices
                 .Select(dto =>
                 {
                     var model = _mapper.Map<DeviceTypeDisplayModel>(dto);
-                    model.IconUrl = _deviceTypeIconStorageService.BuildPublicUrl(model.IconName, model.IconUpdatedAt);
                     return model;
                 })
                 .ToList();
